Functionality for deleting the RiSc in the plugin (#551)
Adds functionality for deleting the RiSc in the plugin. Deletion happens in two different ways, depending on the state of the RiSc:
- If the RiSc has never been published, the RiSc is immediately deleted.
- If the RiSc has been published, the RiSc is deleted on its draft branch (branch is created if it does not exist) and an approval stage is started.

Whenever a RiSc cannot be found on its draft branch, but the branch exists, it is considered marked for deletion. In these cases, the approval flow is shown, with text indicating that the approval is for deletion.

---------

Co-authored-by: Halvard Hummel <halvard.hummel@bekk.no>